challenge-service/requirements.txt
django==2.2.6
djangorestframework==3.10.3
django-rest-auth==0.9.5
django-allauth==0.40.0
mysqlclient==1.4.4
gunicorn==19.9.0
django-cors-headers==3.1.1
sentry-sdk==0.10.2
django-filter==2.2.0
django-oauth-toolkit==1.2.0
drf-tracking==1.5.0
djangorestframework-bulk==0.2.1
django-background-tasks==1.2.5
django-compat==1.0.15
pdfkit==0.6.1
python-magic==0.4.18
xlwt==1.3.0
xmltodict==0.12.0
num2words==0.5.10
xlrd==2.0.1
python-barcode==0.13.1
cryptography==2.8
wheel==0.34.2
pandas
holidays==0.13